refactor: move tracing to own module
