+++ forge/src/lib.rs
@@ -1,5 +1,8 @@
 pub mod decode;
 
+/// Call trace arena, decoding and formatting
+pub mod trace;
+
 /// The Forge test runner
 mod runner;
 pub use runner::{ContractRunner, TestKind, TestKindGas, TestResult};
